tests/pythonscript/requirements.txt
xraylarch
numpy